Drop PGP keys in this site; now separate WKD site
Serving up PGP keys in this repository restricted us to keys supported by
unspecified versions of the GnuPG software inside netlify's builders and has
proven to be fragile such that only one person could support it.

The WKD system allows for using a different hostname for serving keys.  So
openpgpkey.nats.io now exists and holds the PGP keys which used to be here.

Any NATS contributor with an email address in the @nats.io domain can have
their PGP key added to the new site, drop an email to security@nats.io with the
details.